Ajout de application-servlet et welcome.jsp qui n'arrête pas de disparaitre ... Ajout des logs différenciés pour les transactions avec un fichier TRANSACTIONLOG contenant uniquement les transaction réalisées.

+++ webServiceProxiBanque/src/main/java/fr/gtm/pbsi/service/TransactionService.java
@@ -32,7 +32,7 @@ public class TransactionService {
 	@Autowired
 	private ITransactionDao daoTransaction;
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(TransactionService.class);
+	private static final Logger LOGGER = LoggerFactory.getLogger("transactions");
 
 	/**
 	 * Methode post permettant l'insertion en BDD d'une transaction.
@@ -55,8 +55,16 @@ public class TransactionService {
 	 */
 	@DeleteMapping("/{transactionId}")
 	Integer delete(@PathVariable Integer transactionId) {
-		this.daoTransaction.deleteById(transactionId);
-		return 1;
+		if (this.daoTransaction.existsById(transactionId)) {
+			final Optional<Transaction> retour = this.daoTransaction.findById(transactionId);
+			TransactionService.LOGGER.info("Suppression de " + retour.get() + " de la BDD.");
+			this.daoTransaction.deleteById(transactionId);
+			return 1;
+		} else {
+			TransactionService.LOGGER.error("Tentative de suppression d'une transaction échouée car l'ID donné n'est pas trouvable dans la BDD.");
+			return 0;
+		}
+
 	}
 
 	/**
@@ -67,6 +75,7 @@ public class TransactionService {
 	 */
 	@GetMapping({ "", "/" })
 	List<Transaction> readAll() {
+		TransactionService.LOGGER.info("Récupération de la liste de toutes les transactions de ProxiBanque.");
 		return this.daoTransaction.findAll();
 	}
 
@@ -80,8 +89,10 @@ public class TransactionService {
 	Transaction read(@PathVariable Integer transactionId) {
 		final Optional<Transaction> retour = this.daoTransaction.findById(transactionId);
 		if (retour.isPresent()) {
+			TransactionService.LOGGER.info("Récupération de " + retour + ".");
 			return retour.get();
 		} else {
+			TransactionService.LOGGER.error("Tentative de récupération d'une transaction échouée car l'ID donné n'a aucune correspondance en BDD.");
 			final Transaction response = new Transaction();
 			response.setId(0);
 			return response;
@@ -101,8 +112,11 @@ public class TransactionService {
 	@PutMapping("/{transactionId}")
 	Transaction update(@PathVariable Integer transactionId, @RequestBody Transaction transaction) {
 		if (this.daoTransaction.existsById(transactionId)) {
-			return this.daoTransaction.save(transaction);
+			final Transaction retour = this.daoTransaction.save(transaction);
+			TransactionService.LOGGER.info("Modification de " + transaction + " en " + retour + " dans la BDD.");
+			return retour;
 		} else {
+			TransactionService.LOGGER.error("Tentative de modification de " + transaction + " échouée car cet employe n'existe pas en BDD ou l'ID de la requête n'a pas de correspondance en BDD.");
 			final Transaction response = new Transaction();
 			response.setId(0);
 			return response;
